+++ src/components/renderForm.tsx
@@ -79,7 +79,34 @@ const RenderForm = ({
       referoProps.customNavigationCallBack('back', undefined);
     }
   };
-
+  const pauseButtonText = referoProps.customProps?.isMicroweb
+    ? resources?.microwebstep_navigasjon_tilbake_button || 'Tilbake'
+    : displayPreviousButton && isStepView
+      ? resources.previousStep || 'Lagre'
+      : resources.formSave;
+  const onSubmitButtonClicked = referoProps.customProps?.isMicroweb
+    ? methods.handleSubmit(onSubmitReactHookForm, onErrorReactHookForm)
+    : displayNextButton
+      ? methods.handleSubmit(handleNextStep, onErrorReactHookForm)
+      : methods.handleSubmit(onSubmitReactHookForm, onErrorReactHookForm);
+  const cancelButtonText = referoProps.customProps?.isMicroweb
+    ? resources?.microwebstep_navigasjon_avbryt_button || 'Avbryt'
+    : resources.formCancel;
+  const submitButtonText = referoProps.customProps?.isMicroweb
+    ? resources?.microwebstep_navigasjon_neste_button || 'Neste'
+    : displayNextButton && resources.nextStep
+      ? resources.nextStep
+      : resources.formSend;
+  const onPauseButtonClicked = referoProps.customProps?.isMicroweb
+    ? backButtonClicked
+    : isStepView
+      ? displayPauseButtonInStepView
+      : displayPauseButtonInNormalView;
+  const isStepViewProp = referoProps.customProps?.isMicroweb ? false : isStepView;
+  const submitButtonDisabled = referoProps.customProps?.isMicroweb ? undefined : referoProps.submitButtonDisabled;
+  const pauseButtonDisabled = referoProps.customProps?.isMicroweb ? undefined : referoProps.saveButtonDisabled;
+  const onCancelButtonClicked = referoProps.customProps?.isMicroweb ? cancelButtonClicked : referoProps.onCancel;
+  const overrideButtonOrder = referoProps.customProps?.isMicroweb ? buttonOrderMicrowebStep : undefined;
   return (
     <>
       <form onSubmit={methods.handleSubmit(onSubmitReactHookForm, onErrorReactHookForm)}>
@@ -87,42 +114,31 @@ const RenderForm = ({
         {children}
         {!displayValidationSummaryOnTop && <ValidationSummary resources={resources} />}
       </form>
-
-      {referoProps.customProps?.isMicroweb && (
+      {referoProps.renderCustomActionButtons ? (
+        referoProps.renderCustomActionButtons({
+          isStepView,
+          referoProps,
+          displayNextButton,
+          displayPreviousButton,
+          nextStep,
+          previousStep,
+          reactHookFormMethods: methods,
+        })
+      ) : (
         <FormButtons
-          isStepView={false}
-          submitButtonText={resources?.microwebstep_navigasjon_neste_button || 'Neste'}
-          cancelButtonText={resources?.microwebstep_navigasjon_avbryt_button || 'Avbryt'}
-          pauseButtonText={resources?.microwebstep_navigasjon_tilbake_button || 'Tilbake'}
-          onSubmitButtonClicked={methods.handleSubmit(onSubmitReactHookForm, onErrorReactHookForm)}
-          onCancelButtonClicked={cancelButtonClicked}
-          onPauseButtonClicked={backButtonClicked}
+          isStepView={isStepViewProp}
+          submitButtonText={submitButtonText}
+          cancelButtonText={cancelButtonText}
+          pauseButtonText={pauseButtonText}
+          submitButtonDisabled={submitButtonDisabled}
+          pauseButtonDisabled={pauseButtonDisabled}
+          onSubmitButtonClicked={onSubmitButtonClicked}
+          onCancelButtonClicked={onCancelButtonClicked}
+          onPauseButtonClicked={onPauseButtonClicked}
           isAuthorized={isAuthorized}
           loginButton={referoProps.loginButton}
-          overrideButtonOrder={buttonOrderMicrowebStep}
-          isMicrowebStep={true}
-        />
-      )}
-
-      {referoProps.customProps?.isMicroweb === undefined && (
-        <FormButtons
-          isStepView={isStepView}
-          submitButtonText={displayNextButton && resources.nextStep ? resources.nextStep : resources.formSend}
-          cancelButtonText={resources.formCancel}
-          pauseButtonText={displayPreviousButton && isStepView ? resources.previousStep || 'Lagre' : resources.formSave}
-          submitButtonDisabled={referoProps.submitButtonDisabled}
-          pauseButtonDisabled={referoProps.saveButtonDisabled}
-          onSubmitButtonClicked={
-            displayNextButton
-              ? methods.handleSubmit(handleNextStep, onErrorReactHookForm)
-              : methods.handleSubmit(onSubmitReactHookForm, onErrorReactHookForm)
-          }
-          onCancelButtonClicked={(): void => {
-            referoProps.onCancel && referoProps.onCancel();
-          }}
-          onPauseButtonClicked={isStepView ? displayPauseButtonInStepView : displayPauseButtonInNormalView}
-          isAuthorized={isAuthorized}
-          loginButton={referoProps.loginButton}
+          overrideButtonOrder={overrideButtonOrder}
+          isMicrowebStep={!!referoProps.customProps?.isMicroweb}
         />
       )}
     </>